Move providers from app module to core module
