+++ src/app/app.module.ts
@@ -1,38 +1,13 @@
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import { APP_INITIALIZER, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { CookieModule } from 'ngx-cookie';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CoreModule } from './core/core.module';
-import { AuthInterceptor } from './core/interceptors/auth.interceptor';
+import { MainAppModule } from './modules/main-app.module';
 import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
 import { MainLayoutComponent } from './shared/layouts/main-layout/main-layout.component';
-import { AuthService } from './core/services/auth.service';
-import { MainAppModule } from './modules/main-app.module';
 import { PageNotFoundComponent } from './shared/pages/page-not-found/page-not-found.component';
-
-function resourceProviderFactory(authService: AuthService) {
-  return () => {
-    console.log('app init');
-    return new Promise((resolve, reject) => {
-      authService.autoLogIn().subscribe({
-        next: (user) => {
-          console.log('app init user: ', user);
-          authService.user$.next(user);
-          resolve(true);
-        },
-        error: (err) => {
-          console.log('error');
-          console.log('err');
-          authService.logUserOut();
-          resolve(true);
-        },
-      });
-    });
-  };
-}
 
 @NgModule({
   declarations: [
@@ -41,22 +16,8 @@ function resourceProviderFactory(authService: AuthService) {
     AuthLayoutComponent,
     PageNotFoundComponent,
   ],
-  imports: [
-    BrowserModule,
-    CookieModule.forRoot(),
-    CoreModule,
-    MainAppModule,
-    AppRoutingModule,
-  ],
-  providers: [
-    {
-      provide: APP_INITIALIZER,
-      useFactory: resourceProviderFactory,
-      deps: [AuthService],
-      multi: true,
-    },
-    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
-  ],
+  imports: [BrowserModule, CoreModule, MainAppModule, AppRoutingModule],
+  providers: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
