+++ src/app/core/core.module.ts
@@ -1,13 +1,42 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { APP_INITIALIZER, NgModule } from '@angular/core';
+
+import { CookieModule } from 'ngx-cookie';
 import { AuthModule } from './auth/auth.module';
-import { AuthLayoutComponent } from '../shared/layouts/auth-layout/auth-layout.component';
-import { MainLayoutComponent } from '../shared/layouts/main-layout/main-layout.component';
+import { AuthInterceptor } from './interceptors/auth.interceptor';
+import { AuthService } from './services/auth.service';
+
+function resourceProviderFactory(authService: AuthService) {
+  return () => {
+    return new Promise((resolve, reject) => {
+      authService.autoLogIn().subscribe({
+        next: (user) => {
+          authService.user$.next(user);
+          resolve(true);
+        },
+        // TODO Must check for server errors (offline) and do something about it
+        error: (err) => {
+          authService.logUserOut();
+          resolve(true);
+        },
+      });
+    });
+  };
+}
 
 @NgModule({
   declarations: [],
-  imports: [CommonModule, AuthModule],
-  providers: [],
+  imports: [CommonModule, AuthModule, CookieModule.forRoot()],
   exports: [],
+  providers: [
+    {
+      provide: APP_INITIALIZER,
+      useFactory: resourceProviderFactory,
+      deps: [AuthService],
+      multi: true,
+    },
+    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+  ],
 })
 export class CoreModule {}
